pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "archaeohack"
version = "0.1.0"
description = "GUI + predictor for ArchaeoHack Hieroglyph Classifier"
requires-python = ">=3.9"

dependencies = [
    "Pillow",
    "ttkbootstrap",
    "customtkinter",
    "numpy",
    "onnxruntime",
]
